Create License.md
This is the Creative Commons most restrictive License. I may decide to change the License to a less restrictive one in the future. I don't know, as this is the first time I've ever used one, and this is the first thing I have ever done. So, I can say I don't know what I want to do.